Iterative deepening implemented
Iterative deepening is now implemented. Along with the transposition tables, search algorithms are done for now, and next task should be working on move ordering. Piece values were also normalized to the pawn (pawn is now equal to 100 points both in the middle game and the endgame, and rest of piece values were adjusted accordingly).